Fixed last comments, still not added new tests
