+++ test/tests/BuilderTest.java
@@ -1,7 +1,6 @@
 package tests;
 
 import builder.order.Order;
-import builder.order.OrderBuilder;
 import builder.orderitem.OrderItem;
 import org.junit.Before;
 import org.junit.Test;
@@ -14,13 +13,16 @@ import static org.junit.Assert.assertTrue;
 
 public class BuilderTest {
     private Order order;
+    private OrderItem item;
 
     @Before
     public void setUp() throws Exception {
         List<OrderItem> list = new ArrayList<>();
+        item = new OrderItem.OrderItemBuilder().productName("product").price(10.00).build();
+        list.add(item);
         Date creationDate = new Date(2016, 12, 5);
         Date deliveryDate = new Date(2016, 12, 10);
-        order = new Order(new OrderBuilder("name", list).buildCreationDate(creationDate).buildDeliveryDate(deliveryDate));
+        order = new Order.OrderBuilder().customerName("name").items(list).creationDate(creationDate).deliveryDate(deliveryDate).build();
     }
 
     private boolean isOnTime(Order order, int days) {
